fix(bridge): 浮点绑定改 Qv::Double 防 f32 截断；reqwest client 构建失败 fail-fast
- query.rs to_qv: Qv::Float(f as f32) → Qv::Double(f)，金额/高精度小数不再丢精度
- mod.rs: client builder unwrap_or_default → expect（Default 兜底本身会 panic，改显式 fail-fast）

unix@vip.qq.com ai

+++ src/bridge/query.rs
@@ -94,7 +94,7 @@ fn to_qv(v: &Value) -> Qv {
             if let Some(i) = n.as_i64() {
                 Qv::BigInt(Some(i))
             } else if let Some(f) = n.as_f64() {
-                Qv::Float(Some(f as f32))
+                Qv::Double(Some(f))
             } else {
                 Qv::String(None)
             }
